working on commands, need to refactor everything (current design sucks in rust)

+++ src/command.rs
@@ -6,12 +6,12 @@ pub enum Command {
         location: String,
     },
     Remove {
-    	id: i32,
+    	id: usize,
     	location: String,
     },
     Move {
         from: String,
-        id: i32,
+        id: usize,
         to: String
     }
 }
